+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed missing headers from some commands in the cli
+
+### Added
+- Added more documentation
 
 ## [0.3.5] - 2024-08-11
 ### Fixed
@@ -154,4 +159,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.1.2]: https://github.com/adnaniazi/capfinder/compare/0.1.1...0.1.2
 [0.1.1]: https://pypi.org/manage/project/capfinder/release/0.1.1/
 [0.1.0]: https://pypi.org/manage/project/capfinder/release/0.1.0/
-