+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.6] - 2024-08-13
 ### Fixed
 - Fixed missing headers from some commands in the cli
 
@@ -133,7 +135,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Basic skeleton of the package and tested it
 
-[Unreleased]: https://github.com/adnaniazi/capfinder/compare/0.3.5...master
+[Unreleased]: https://github.com/adnaniazi/capfinder/compare/0.3.6...master
+[0.3.6]: https://github.com/adnaniazi/capfinder/compare/0.3.5...0.3.6
 [0.3.5]: https://github.com/adnaniazi/capfinder/compare/0.3.4...0.3.5
 [0.3.4]: https://github.com/adnaniazi/capfinder/compare/0.3.3...0.3.4
 [0.3.3]: https://github.com/adnaniazi/capfinder/compare/0.3.2...0.3.3
@@ -160,3 +163,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.1.2]: https://github.com/adnaniazi/capfinder/compare/0.1.1...0.1.2
 [0.1.1]: https://pypi.org/manage/project/capfinder/release/0.1.1/
 [0.1.0]: https://pypi.org/manage/project/capfinder/release/0.1.0/
+